doge-response/requirements.txt
google-api-python-client==1.7.11
flask==1.1.1
slackeventsapi==2.1.0
google-cloud-vision==1.0.0
pillow==7.0.0
google-cloud-storage==1.26.0
google-cloud-pubsub==1.4.2